fix async events converte cluod points

+++ service_subsampling/main.py
@@ -30,14 +30,19 @@ async def upload_file(file: UploadFile = File(...), factor: int = 10):
 
     # Call the subsample_point_cloud function
     subsample_point_cloud(temp_input_file, temp_output_file_las, factor)
-    run_potree_converter(temp_output_file_las, temp_output_file)
-    zip_folder(temp_output_file, temp_output_zip)
-    upload_to_blob_storage(connection_string, container_name, temp_output_zip, f"potree_{file.filename[:-4]}.zip")
-
-    shutil.rmtree("tmp")
-    os.makedirs("tmp")
     
-    return JSONResponse(content={"filename": f"reduced_{file.filename}", "content_type": file.content_type})
+    success = await run_potree_converter(temp_output_file_las, temp_output_file)
+    if success:
+        zip_folder(temp_output_file, temp_output_zip)
+        upload_to_blob_storage(connection_string, container_name, temp_output_zip, f"potree_{file.filename[:-4]}.zip")
+        shutil.rmtree("tmp")
+        os.makedirs("tmp")
+        return JSONResponse(content={"filename": f"reduced_{file.filename}", "content_type": file.content_type})
+        
+    else:
+        raise HTTPException(status_code=500, detail="Error running Potree converter.")
+    
+
 
 
 if __name__ == "__main__":
